requirements.txt
aioredis==1.3.1
alabaster==0.7.8
altgraph==0.17
anyascii==0.1.7
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
arrow==0.17.0
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
auto-py-to-exe==2.7.11
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.2.1
blessed==1.17.12
blinker==1.4
bootstrap4==0.1.0
boto3==1.16.4
botocore==1.19.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2019.11.28
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.1
command-not-found==0.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
crispy-forms-materialize==0.2
cryptography==3.1.1
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.10.0
daphne==3.0.1
day==0.0.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.1
django-activity-stream==0.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==3.0.9
django-extra-fields==3.0.2
django-filter==2.4.0
django-htmlmin==0.11.0
django-humanize==0.1.2
django-modelcluster==5.1
django-multiselectfield==0.1.12
django-mysql==3.10.0
django-picklefield==3.0.1
django-q==1.3.4
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-storages==1.10.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.16
draftjs-exporter==2.1.7
drf-api-tracking==1.7.7
drf-yasg==1.20.0
drf-yasg2==1.19.3
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
environ==1.0
et-xmlfile==1.0.1
Faker==0.9.1
fastapi==0.63.0
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
gevent==20.12.1
gevent-websocket==0.10.1
gitsome==0.8.0
graphviz==0.14.2
greenlet==1.0a1
gunicorn==20.0.4
h11==0.11.0
hiredis==1.1.0
html5lib==1.0.1
httplib2==0.14.0
humanize==3.1.0
hyperlink==20.0.1
idna==2.8
imagesize==1.2.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
ipaddr==2.2.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kazam==1.5.3
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.1
l18n==2020.6.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
livereload==2.6.3
lockfile==0.12.2
login==0.0.6
louis==3.12.0
lunr==0.5.8
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mixer==6.1.3
model-bakery==1.2.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.1
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
nltk==3.5
notebook==6.1.5
numpy==1.19.4
numpydoc==0.7.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.3
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.15
pluggy==0.13.1
ply==3.11
prettytable==2.0.0
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
propose==0.1.0a0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.7.3
Pygments==2.7.2
PyGObject==3.36.0
PyHamcrest==2.0.2
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
Pyphen==0.10.0
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.2.0
pytest-django==4.1.0
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-xlib==0.23
pytoml==0.1.21
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
roman==2.0.0
rope==0.18.0
rply==0.7.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.17.2
six==1.15.0
soft-delete==0.2.2
soupsieve==2.0.1
Sphinx==1.8.5
sqlparse==0.4.1
starlette==0.13.6
swagger-spec-validator==2.7.3
system-service==0.3
systemd-python==234
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.2
threadpoolctl==2.1.0
tinycss2==1.1.0
today==0.2
toml==0.10.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.1
unattended-upgrades==0.1
Unidecode==1.1.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.12.2
virtualenv==20.4.0
virtualenv-clone==0.5.4
wadllib==1.3.3
wagtail==2.11.2
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
websocket-client==0.57.0
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
Willow==1.4
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
xonsh==0.9.13
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.1.2
